requirements.txt
numpy
tensorflow
python-chess
zstandard
